refactor folders of notifications feature
