+++ lib/features/admin/home_admin/data/drawer_items_list.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:my_store/features/admin/categories/presentation/views/categories_screen.dart';
 import 'package:my_store/features/admin/dashboard/views/presentation/dashboard_screen.dart';
 import 'package:my_store/features/admin/home_admin/data/models/drawer_item_model.dart';
-import 'package:my_store/features/admin/notifications/presentation/notifications_screen.dart';
+import 'package:my_store/features/admin/notifications/presentation/views/notifications_screen.dart';
 import 'package:my_store/features/admin/products/presentation/views/products_screen.dart';
 import 'package:my_store/features/admin/users/presentation/views/users_screen.dart';
