+++ lib/features/admin/notifications/presentation/views/notifications_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:my_store/core/common/widgets/custom_admin_appbar.dart';
-import 'package:my_store/features/admin/notifications/presentation/widgets/add_notification_item.dart';
-import 'package:my_store/features/admin/notifications/presentation/widgets/notifications_list_view.dart';
+import 'package:my_store/features/admin/notifications/presentation/views/widgets/add_notification_item.dart';
+import 'package:my_store/features/admin/notifications/presentation/views/widgets/notifications_list_view.dart';
 
 class NotificationsScreen extends StatelessWidget {
   const NotificationsScreen({super.key});
