+++ lib/features/admin/notifications/presentation/views/widgets/add_notification_bottom_sheet.dart
@@ -3,10 +3,10 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:my_store/core/common/widgets/custom_text_field.dart';
 import 'package:my_store/core/extensions/theme_context.dart';
 import 'package:my_store/core/style/fonts/font_weight_helper.dart';
-import 'package:my_store/features/admin/notifications/presentation/widgets/edit_notification_button.dart';
+import 'package:my_store/features/admin/notifications/presentation/views/widgets/add_notification_button.dart';
 
-class EditNotificationBottomSheet extends StatelessWidget {
-  const EditNotificationBottomSheet({super.key});
+class AddNotificationBottomSheet extends StatelessWidget {
+  const AddNotificationBottomSheet({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -18,7 +18,7 @@ class EditNotificationBottomSheet extends StatelessWidget {
           children: [
             Align(
               child: Text(
-                'Edit Notification',
+                'Add Notification',
                 style: context.textStyle.copyWith(
                   fontSize: 20.sp,
                   fontWeight: FontWeightHelper.bold,
@@ -27,7 +27,7 @@ class EditNotificationBottomSheet extends StatelessWidget {
             ),
             SizedBox(height: 20.h),
             Text(
-              'Edit the notification title',
+              'Enter the notification title',
               style: context.textStyle.copyWith(
                 fontSize: 16.sp,
                 fontWeight: FontWeightHelper.medium,
@@ -47,7 +47,7 @@ class EditNotificationBottomSheet extends StatelessWidget {
             ),
             SizedBox(height: 20.h),
             Text(
-              'Edit the notification body',
+              'Enter the notification body',
               style: context.textStyle.copyWith(
                 fontSize: 16.sp,
                 fontWeight: FontWeightHelper.medium,
@@ -67,7 +67,7 @@ class EditNotificationBottomSheet extends StatelessWidget {
             ),
             SizedBox(height: 20.h),
             Text(
-              'Edit the product id',
+              'Enter the product id',
               style: context.textStyle.copyWith(
                 fontSize: 16.sp,
                 fontWeight: FontWeightHelper.medium,
@@ -86,7 +86,7 @@ class EditNotificationBottomSheet extends StatelessWidget {
               },
             ),
             SizedBox(height: 20.h),
-            const EditNotificationButton(),
+            const AddNotificationButton(),
           ],
         ),
       ),
