+++ lib/features/admin/notifications/presentation/views/widgets/add_notification_item.dart
@@ -4,7 +4,7 @@ import 'package:my_store/core/common/widgets/custom_linear_button.dart';
 import 'package:my_store/core/extensions/theme_context.dart';
 import 'package:my_store/core/style/colors/colors_dark.dart';
 import 'package:my_store/core/style/fonts/font_weight_helper.dart';
-import 'package:my_store/features/admin/notifications/presentation/widgets/add_notification_bottom_sheet.dart';
+import 'package:my_store/features/admin/notifications/presentation/views/widgets/add_notification_bottom_sheet.dart';
 
 class AddNotificationItem extends StatelessWidget {
   const AddNotificationItem({
